chore: sep. DropdownText comp.
